Write a dev Dockerfile for K8s unit tests
Craft it specifically to cache as much of the environment and dependency
list as possible so developers can have a rapid feedback loop.

I don't want people confused about the purpose of this Dockerfile. It's
only meant to be used as a platform on which developers can easily run
presubmit checks. To this end, I've added a `docker-test` target to the
Makefile that runs a subset of the presumit checks (using the current
tree) in a Docker container.

I can't go farther down this path until I find a way to deal with the
integration tests that currently expect to use Docker for some of their
work.