applyManifestToResource: directly walk the JSON
Previous behavior was to do a complicated Unmarshal/Marshal + typecast
transform which did not mutate the original object.

This method I am using now does mutate the original object, but for some
reason the final object that is printed is the original version. I'm
confused about this behavior and I will debug it in the next commit.